requirements.txt
amqp==5.2.0
asgiref==3.7.2
asyncio==3.4.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.0
cryptography==41.0.7
daphne==4.0.0
Django==4.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
flake8==7.0.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kombu==5.3.4
Markdown==3.5.2
mccabe==0.7.0
msgpack==1.0.7
packaging==23.2
pluggy==1.3.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
pyOpenSSL==23.3.0
pytest==7.4.4
pytest-django==4.7.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.1
requests==2.31.0
rpds-py==0.16.2
service-identity==23.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.1